Updated esl-alimerge test and man page to include --rfonly option, added in previous revision (r460).
git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@461 db3e70e4-12e1-0310-b23f-81a6e3d09ba7

+++ miniapps/esl-alimerge.itest.pl
@@ -39,7 +39,7 @@ print LISTFILE "$tmppfx.2\n";
 close LISTFILE;
 
 $output = `$eslalimerge -h`;
-if ($? != 0)                                     { die "FAIL: esl-alimerge failed unexpectedly"; }
+if ($? != 0)                                         { die "FAIL: esl-alimerge failed unexpectedly"; }
 if ($output !~ /Usage: esl-alimerge/)                { die "FAIL: help output not right"; }
 
 $output = `$eslalimerge --rna $tmppfx.1 $tmppfx.2 2>&1`;
@@ -102,6 +102,12 @@ $output = `$eslalimerge --rna --outformat afa $tmppfx.1 $tmppfx.2 2>&1`;
 if ($? != 0)                                                                              { die "FAIL: esl-alimerge failed unexpectedly"; }
 if ($output !~ /aaAAAA\.A\.\.AAA\.\.\.\.\.Cc\.cCC\.\.CCCC\.\.\.\.C\.\.G\.\.GG\.\.GGgggg/) { die "FAIL: alignments merged incorrectly"; }
 if ($output !~ /\.\.AAAA\.AaaAAAaccccC\.\.\.\-CccCCCCccccccgGggGGggGGg\.\.\./)            { die "FAIL: alignments merged incorrectly"; }
+
+$output = `$eslalimerge --rna --rfonly $tmppfx.1 $tmppfx.2 2>&1`;
+if ($? != 0)                              { die "FAIL: esl-alimerge failed unexpectedly"; }
+if ($output !~ /sequence 3 is the best/)  { die "FAIL: alignments merged incorrectly"; }
+if ($output !~ /AAAAAAAACCCCCCCCGGGGG/)   { die "FAIL: alignments merged incorrectly"; }
+if ($output !~ /AAAAAAAAC\-CCCCCcGGGGG/)  { die "FAIL: alignments merged incorrectly"; }
 
 print "ok\n"; 
 unlink "$tmppfx.1";
